userTask/feature: added recognition of spoken words task, redesigned userTask class structure
