+++ src/js/game.js
@@ -1,4 +1,6 @@
 import GameInterface from './modules/gameInterface/gameInterface';
+import UserTask from './modules/userTask/userTask';
 
 new GameInterface();
 
+const x = new UserTask('speech');
