+++ src/js/modules/userTask/userTask.js
@@ -1,7 +1,10 @@
 import _ from 'lodash';
-import $ from 'jquery';
-import 'jquery-ui-sortable-npm';
 import taskConfig from '../../taskConfig';
+import generateAuditionTask from './tasks/audition';
+import generateCodeTask from './tasks/code';
+import generateSequenceTask from './tasks/sequence';
+import generateTranslateTask from './tasks/translate';
+import generateSpeechTask from './tasks/speech';
 
 /*eslint no-case-declarations: 0*/
 
@@ -18,12 +21,10 @@ export default class UserTask {
 			function clickHandler(e) {
 				if (e.target === submit) {
 					const answer = that.checkAnswer(that.taskObject);
-
 					if (answer !== undefined) {
 						resolve(answer);
 						document.querySelector('.modal-overlay').remove();
 					}
-
 				}
 			}
 
@@ -42,7 +43,7 @@ export default class UserTask {
 		});
 	}
 	checkTaskType(type) {
-		if (typeof type === 'string' && type === 'code' || type === 'translate' || type === 'sequence' || type === 'audition') {
+		if (typeof type === 'string' && type === 'code' || type === 'translate' || type === 'sequence' || type === 'audition' || type === 'speech') {
 			return type;
 		} else {
 			return 'code';
@@ -65,236 +66,29 @@ export default class UserTask {
 			return taskConfig.sequence[_.random(0, taskConfig.sequence.length - 1)];
 		case 'audition':
 			return taskConfig.audition[_.random(0, taskConfig.audition.length - 1)];
+		case 'speech':
+			return taskConfig.speech[_.random(0, taskConfig.speech.length - 1)];
 		default:
 			return taskConfig.code[_.random(0, taskConfig.code.length - 1)];
 		}
 	}
 
-	generateCodeTask(task) {
-		const modalOverlay = document.createElement('div');
-		modalOverlay.className = 'modal-overlay';
-		modalOverlay.setAttribute('id', 'modal-overlay');
-		const taskModal = document.createElement('div');
-		taskModal.className = 'task-modal';
-		modalOverlay.appendChild(taskModal);
-		const title = document.createElement('h2');
-		title.className = 'task-modal-title';
-		taskModal.appendChild(title);
-		const question = document.createElement('div');
-		question.className = 'task-modal-question';
-		taskModal.appendChild(question);
-
-		const answer = document.createElement('div');
-		answer.className = 'task-modal-answer';
-		const answerTitle = document.createElement('p');
-		answerTitle.className = 'task-modal-answer-title';
-		answerTitle.textContent = 'Ваш ответ:';
-		answer.appendChild(answerTitle);
-		const answerVariantsList = document.createElement('ul');
-		answerVariantsList.className = 'task-modal-anwer-variants';
-
-		title.textContent = 'Задача по JavaScript';
-		question.innerHTML = task[0];
-
-		for (let i = 0; i < task[1].length; i++) {
-			const answerVariantItem = document.createElement('li');
-			const label = document.createElement('label');
-			answerVariantItem.appendChild(label);
-			const input = document.createElement('input');
-			input.setAttribute('type', 'radio');
-			input.setAttribute('name', 'answer');
-			input.setAttribute('value', task[1][i]);
-			label.appendChild(input);
-			const inputText = document.createElement('code');
-			inputText.textContent = task[1][i];
-			label.appendChild(inputText);
-			answerVariantsList.appendChild(answerVariantItem);
-			answer.appendChild(answerVariantsList);
-			taskModal.appendChild(answer);
-		}
-
-		const button = document.createElement('button');
-		button.className = 'task-modal-btn';
-		button.setAttribute('id', 'task-answer');
-		button.setAttribute('type', 'submit');
-		button.textContent = 'sumbit';
-		taskModal.appendChild(button);
-		document.getElementById('game-container').appendChild(modalOverlay);
-	}
-
-	generateTranslateTask(task) {
-		const modalOverlay = document.createElement('div');
-		modalOverlay.className = 'modal-overlay';
-		modalOverlay.setAttribute('id', 'modal-overlay');
-		const taskModal = document.createElement('div');
-		taskModal.className = 'task-modal';
-		modalOverlay.appendChild(taskModal);
-		const title = document.createElement('h2');
-		title.className = 'task-modal-title';
-		taskModal.appendChild(title);
-		const question = document.createElement('div');
-		question.className = 'task-modal-question';
-		taskModal.appendChild(question);
-
-		const answer = document.createElement('div');
-		answer.className = 'task-modal-answer';
-		const answerTitle = document.createElement('p');
-		answerTitle.className = 'task-modal-answer-title';
-		answerTitle.textContent = 'Ваш ответ:';
-		answer.appendChild(answerTitle);
-		const answerVariantsList = document.createElement('ul');
-		answerVariantsList.className = 'task-modal-anwer-variants';
-
-		title.textContent = 'Перевод слова';
-		question.innerHTML = `Переведите с английского на русский слово: <code><b>${task[0]}</b></code>`;
-
-		const answerVariantItem = document.createElement('li');
-		const label = document.createElement('label');
-		answerVariantItem.appendChild(label);
-		const input = document.createElement('input');
-		input.setAttribute('type', 'text');
-		input.setAttribute('name', 'answer');
-		label.appendChild(input);
-		answerVariantsList.appendChild(answerVariantItem);
-		answer.appendChild(answerVariantsList);
-		taskModal.appendChild(answer);
-
-		const button = document.createElement('button');
-		button.className = 'task-modal-btn';
-		button.setAttribute('id', 'task-answer');
-		button.setAttribute('type', 'submit');
-		button.textContent = 'sumbit';
-		taskModal.appendChild(button);
-		document.getElementById('game-container').appendChild(modalOverlay);
-	}
-
-	generateSequenceTask(task) {
-		const modalOverlay = document.createElement('div');
-		modalOverlay.className = 'modal-overlay';
-		modalOverlay.setAttribute('id', 'modal-overlay');
-		const taskModal = document.createElement('div');
-		taskModal.className = 'task-modal';
-		modalOverlay.appendChild(taskModal);
-		const title = document.createElement('h2');
-		title.className = 'task-modal-title';
-		taskModal.appendChild(title);
-		const question = document.createElement('div');
-		question.className = 'task-modal-question';
-		taskModal.appendChild(question);
-
-		const answer = document.createElement('div');
-		answer.className = 'task-modal-answer';
-		const answerTitle = document.createElement('p');
-		answerTitle.className = 'task-modal-answer-title';
-		answerTitle.textContent = 'Ваш ответ:';
-		answer.appendChild(answerTitle);
-		const answerVariantsList = document.createElement('ul');
-		answerVariantsList.className = 'task-modal-anwer-variants';
-
-		title.textContent = 'Правильная последовательность';
-		question.innerHTML = 'Расположите в правильной последовательности следующий код:';
-		const taskShuffle = _.shuffle(task);
-		for (let i = 0; i < task.length; i++) {
-			const answerVariantItem = document.createElement('li');
-			const pre = document.createElement('pre');
-			const inputText = document.createElement('code');
-			pre.appendChild(inputText);
-			inputText.textContent = taskShuffle[i];
-			answerVariantsList.classList.add('task-modal-answer-sequence');
-			answerVariantItem.appendChild(pre);
-			answerVariantsList.appendChild(answerVariantItem);
-			answer.appendChild(answerVariantsList);
-			taskModal.appendChild(answer);
-		}
-
-		const button = document.createElement('button');
-		button.className = 'task-modal-btn';
-		button.setAttribute('id', 'task-answer');
-		button.setAttribute('type', 'submit');
-		button.textContent = 'sumbit';
-		taskModal.appendChild(button);
-		document.getElementById('game-container').appendChild(modalOverlay);
-		$('.task-modal-anwer-variants').sortable();
-	}
-
-	generateAuditionTask(task) {
-		const modalOverlay = document.createElement('div');
-		modalOverlay.className = 'modal-overlay';
-		modalOverlay.setAttribute('id', 'modal-overlay');
-		const taskModal = document.createElement('div');
-		taskModal.className = 'task-modal';
-		modalOverlay.appendChild(taskModal);
-		const title = document.createElement('h2');
-		title.className = 'task-modal-title';
-		taskModal.appendChild(title);
-		const question = document.createElement('div');
-		question.className = 'task-modal-question';
-		taskModal.appendChild(question);
-
-		const answer = document.createElement('div');
-		answer.className = 'task-modal-answer';
-		const answerTitle = document.createElement('p');
-		answerTitle.className = 'task-modal-answer-title';
-		answerTitle.textContent = 'Ваш ответ:';
-		answer.appendChild(answerTitle);
-		const answerVariantsList = document.createElement('ul');
-		answerVariantsList.className = 'task-modal-anwer-variants';
-		title.textContent = 'Аудирование';
-		question.innerHTML = 'Ваша задача написать воспроизведенное слово. Нажмите на конопку, чтобы воспроизвести.';
-		const btn = document.createElement('button');
-		btn.className = 'btn';
-		btn.textContent = 'Воспроизвести';
-
-		const awaitVoices = new Promise(done =>
-			window.speechSynthesis.onvoiceschanged = done);
-		const speakTaskWord = () => {
-			awaitVoices.then(() => {
-				const synth = window.speechSynthesis;
-				const speech = new SpeechSynthesisUtterance(task);
-				let voices = [];
-				voices = synth.getVoices();
-				speech.voice = voices[8];
-				speech.rate = 0.8;
-				synth.speak(speech);
-			});
-		};
-
-		btn.onclick = speakTaskWord;
-		question.appendChild(btn);
-
-		const answerVariantItem = document.createElement('li');
-		const label = document.createElement('label');
-		answerVariantItem.appendChild(label);
-		const input = document.createElement('input');
-		input.setAttribute('type', 'text');
-		input.setAttribute('name', 'answer');
-		label.appendChild(input);
-		answerVariantsList.appendChild(answerVariantItem);
-		answer.appendChild(answerVariantsList);
-		taskModal.appendChild(answer);
-
-		const button = document.createElement('button');
-		button.className = 'task-modal-btn';
-		button.setAttribute('id', 'task-answer');
-		button.setAttribute('type', 'submit');
-		button.textContent = 'sumbit';
-		taskModal.appendChild(button);
-		document.getElementById('game-container').appendChild(modalOverlay);
-	}
-
 	renderTask(task) {
 		switch (this.type) {
 		case 'code':
-			this.generateCodeTask(task);
+			generateCodeTask(task);
 			break;
 		case 'translate':
-			this.generateTranslateTask(task);
+			generateTranslateTask(task);
 			break;
 		case 'sequence':
-			this.generateSequenceTask(task);
+			generateSequenceTask(task);
 			break;
 		case 'audition':
-			this.generateAuditionTask(task);
+			generateAuditionTask(task);
+			break;
+		case 'speech':
+			generateSpeechTask(task);
 			break;
 		}
 	}
@@ -348,8 +142,16 @@ export default class UserTask {
 				}
 			}
 			break;
+		case 'speech':
+			userAnswer = document.querySelector('input[name=answer]');
+			if (userAnswer !== null) {
+				if (userAnswer.value.trim().toLowerCase() === task[0]) {
+					return true;
+				} else {
+					return false;
+				}
+			}
+			break;
 		}
-
 	}
-
 }
